Use alternate implementation sytle based on review feedback.
