+++ sdk/storage_blobs/src/blob/operations/delete_blob_snapshot.rs
@@ -1,4 +1,3 @@
-use crate::blob::Snapshot;
 use crate::{blob::operations::DeleteBlobResponse, prelude::*};
 use azure_core::{headers::Headers, prelude::*};
