+++ sdk/storage_blobs/src/clients/blob_client.rs
@@ -1,7 +1,6 @@
-use crate::blob::Snapshot;
 use crate::{
     blob::operations::*,
-    options::{BA512Range, Tags},
+    options::{BA512Range, Snapshot, Tags},
     prelude::*,
 };
 use azure_core::{
